% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_par_cor.R
\name{bound_dirichlet}
\alias{bound_dirichlet}
\title{Dirichlet parameter bound function}
\usage{
bound_dirichlet(theta)
}
\arguments{
\item{theta}{point to check}
}
\value{
Boolean specifying whether theta lies in the interior of the simplex.
}
\description{
Function calculating membership of a point in the interior of the simplex.
}
\examples{
theta_ext <- runif(100)
theta_ins <- runif(100) / 100
bound_dirichlet(theta_ext)
bound_dirichlet(theta_ins)
}
\keyword{internal}
