% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_quantile.R
\name{trunc_quantile}
\alias{trunc_quantile}
\title{Truncated normal quantiles}
\usage{
trunc_quantile(p, ratio)
}
\arguments{
\item{p}{Percentile}

\item{ratio}{Ratio of standard error to point estimate
(midpoint of confidence interval)}
}
\value{
Truncated normal quantile
}
\description{
This function computes quantiles of the truncated normal distribution
   for calculating THAMES confidence intervals.
}
\keyword{internal}
